fix: implement bulk refresh for Twitter submissions and refactor logic into separate action
